# twitter-go-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance Go backend for the Chanom Twitter clone.",
        "title": "Twitter Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/chanombude/twitter-go-api"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tweets": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Post a new tweet with optional text and media.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tweets"
                ],
                "summary": "Create Tweet",
                "parameters": [
                    {
                        "description": "Tweet content and media",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.createTweetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.TweetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tweets/{id}": {
            "get": {
                "description": "Fetch a single tweet by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tweets"
                ],
                "summary": "Get Tweet",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tweet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.TweetResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a tweet by ID. Only the author can delete their own tweet.",
                "tags": [
                    "Tweets"
                ],
                "summary": "Delete Tweet",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tweet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tweets/{id}/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a like to a tweet.",
                "tags": [
                    "Tweets"
                ],
                "summary": "Like Tweet",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tweet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a like from a tweet.",
                "tags": [
                    "Tweets"
                ],
                "summary": "Unlike Tweet",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tweet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tweets/{id}/replies": {
            "get": {
                "description": "Get a paginated list of replies to a specific tweet.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tweets"
                ],
                "summary": "Get Tweet Replies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tweet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PageResponse-server_TweetResponse"
                        }
                    }
                }
            }
        },
        "/tweets/{id}/retweet": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Share a tweet on your own profile.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tweets"
                ],
                "summary": "Retweet",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tweet ID to retweet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.TweetResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a previously shared retweet.",
                "tags": [
                    "Tweets"
                ],
                "summary": "Undo Retweet",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Original Tweet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/uploads/presign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a temporary presigned URL to upload a file directly to S3.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Uploads"
                ],
                "summary": "Presign Upload URL",
                "parameters": [
                    {
                        "description": "Upload request details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.presignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "presignedUrl and objectKey",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperr.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apperr.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        },
...
        "server.PageResponse-server_TweetResponse": {
            "type": "object",
            "properties": {
                "hasNext": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.TweetResponse"
                    }
                },
                "nextCursor": {
                    "type": "string"
                }
            }
        },
...
        "server.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "server.TweetResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isLiked": {
                    "type": "boolean"
                },
                "isRetweeted": {
                    "type": "boolean"
                },
                "likeCount": {
                    "type": "integer"
                },
                "mediaType": {
                    "type": "string"
                },
                "mediaUrl": {
                    "type": "string"
                },
                "replyCount": {
                    "type": "integer"
                },
                "replyToTweetId": {
                    "type": "integer"
                },
                "replyToUsername": {
                    "type": "string"
                },
                "retweetCount": {
                    "type": "integer"
                },
                "retweetedTweet": {
                    "$ref": "#/definitions/server.TweetResponse"
                },
                "user": {
                    "$ref": "#/definitions/server.UserResponse"
                }
            }
        },
        "server.UserResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "bannerUrl": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "followersCount": {
                    "type": "integer"
                },
                "followingCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isFollowing": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "server.createTweetRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 280
                },
                "mediaKey": {
                    "type": "string"
                },
                "mediaType": {
                    "type": "string",
                    "enum": [
                        "IMAGE",
                        "VIDEO"
                    ]
                },
                "parentId": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "server.presignRequest": {
            "type": "object",
            "required": [
                "contentType",
                "filename",
                "folder"
            ],
            "properties": {
                "contentLength": {
                    "type": "integer",
                    "minimum": 1
                },
                "contentType": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "folder": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer ' followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}